fix(loader): + HTTP tool discovery branch in load_into_with_options (Critical #1)
Wires tool_discovery=Http into the runtime-startup scan. Each catalog entry
becomes a DoraToolBridge with the resolved safety_tier (catalog > tool_overrides
> manifest required_safety_tier). Closes the gap where HTTP-discovered tools
only existed during install and disappeared on next process restart.

- DoraBridge moved from octos-dora-mcp into octos-agent::tools::dora_bridge
  to resolve the dep cycle (octos-dora-mcp now a thin re-export shell).
- PluginLoader::load_into{,_with_work_dir,_with_options} are now async.
- Pipeline handler's plugin cache moved from std::sync::OnceLock to
  tokio::sync::OnceCell so the async loader threads through OnceCell::get_or_init.
- All production callers + ~15 in-file tests updated to .await
  (chat.rs, profile.rs, gateway_runtime.rs, profile_factory.rs, spawn.rs,
  handler.rs, plugin_load_cache.rs).
- 3 new integration tests in runtime_http_discovery cover catalog tier,
  override tier, manifest fallback.
