+++ crates/octos-dora-mcp/src/lib.rs
@@ -1,314 +1,11 @@
-//! Dora-RS to MCP tool bridge for octos.
+//! Compatibility re-export crate.
 //!
-//! Wraps a [`DoraToolMapping`] (a config-defined link from a Dora node output
-//! to an MCP tool name) so the agent's [`Tool`] machinery can dispatch the
-//! request like any other tool. The bridge also registers the tool's
-//! required safety tier in the global [`RobotToolRegistry`] so the existing
-//! `group:robot:<tier>` `ToolPolicy` machinery actually applies — without
-//! that registration the tier metadata would be decorative.
+//! The canonical home for the dora bridge types is now
+//! [`octos_agent::tools::dora_bridge`]. This crate is preserved for the
+//! older `octos-dora-mcp` rename and forwards everything via re-export.
+//!
+//! New code should depend on `octos-agent` directly.
 
-use async_trait::async_trait;
-use octos_agent::tools::ConcurrencyClass;
-use octos_agent::tools::robot_groups::{self, RobotToolRegistry};
-use octos_agent::{SafetyTier, Tool, ToolResult};
-use serde::{Deserialize, Serialize};
-
-pub mod config;
-pub use config::BridgeConfig;
-
-/// Mapping from a catalog entry to an MCP tool exposed via HTTP bridge.
-///
-/// The post-SPEC-V1 1.1 shape: tools are dispatched via `POST <bridge_base_url>/tools/<tool_name>`
-/// and the bridge handles all Dora wiring. The legacy `dora_node_id` /
-/// `dora_output_id` fields remain optional for backward-compatible config
-/// files but are ignored by the HTTP transport.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct DoraToolMapping {
-    /// MCP tool name exposed to the agent.
-    pub tool_name: String,
-    /// Description for the LLM.
-    #[serde(default)]
-    pub description: String,
-    /// Loopback HTTP base URL of the local bridge process (`http://127.0.0.1:PORT`).
-    /// Required for execution; validated against loopback at bridge build time.
-    #[serde(default)]
-    pub bridge_base_url: String,
-    /// Required safety tier for this tool. Strict: unknown tiers fail
-    /// `BridgeConfig::from_json` rather than silently defaulting to `Observe`.
-    #[serde(default)]
-    pub safety_tier: SafetyTier,
-    /// Optional explicit override of the bridge tool's concurrency class.
-    /// When `None`, falls back to the per-tier default in
-    /// `Tool::concurrency_class` (Observe→Safe, anything else→Exclusive).
-    #[serde(default)]
-    pub concurrency_class: Option<ConcurrencyClass>,
-    /// Optional JSON Schema for the tool's input. When `None`, a permissive
-    /// `{"type":"object","additionalProperties":true}` is used.
-    #[serde(default)]
-    pub input_schema: Option<serde_json::Value>,
-    /// Timeout in seconds for the tool call.
-    #[serde(default = "default_timeout")]
-    pub timeout_secs: u64,
-    /// DEPRECATED — kept for older config compatibility but ignored by the
-    /// HTTP-transport implementation. Will be removed in a future release.
-    #[serde(default)]
-    pub dora_node_id: Option<String>,
-    /// DEPRECATED — see `dora_node_id`.
-    #[serde(default)]
-    pub dora_output_id: Option<String>,
-}
-
-fn default_timeout() -> u64 {
-    30
-}
-
-/// A bridge that wraps a [`DoraToolMapping`] as an MCP-compatible [`Tool`].
-///
-/// Transport is delegated to [`HttpTool`]; this layer owns safety_tier
-/// metadata + `RobotToolRegistry` registration so the existing
-/// `group:robot:<tier>` `ToolPolicy` machinery applies.
-pub struct DoraToolBridge {
-    mapping: DoraToolMapping,
-    inner: octos_agent::tools::HttpTool,
-}
-
-impl DoraToolBridge {
-    /// Construct a bridge. Fails if `mapping.bridge_base_url` is not a
-    /// loopback HTTP URL (delegated to `HttpTool::new`).
-    pub fn new(mapping: DoraToolMapping) -> eyre::Result<Self> {
-        let schema = mapping
-            .input_schema
-            .clone()
-            .unwrap_or_else(|| serde_json::json!({"type": "object", "additionalProperties": true}));
-        let inner = octos_agent::tools::HttpTool::new(
-            mapping.tool_name.clone(),
-            mapping.description.clone(),
-            mapping.bridge_base_url.clone(),
-            schema,
-        )?;
-        Ok(Self { mapping, inner })
-    }
-
-    /// Return a reference to the underlying mapping.
-    pub fn mapping(&self) -> &DoraToolMapping {
-        &self.mapping
-    }
-
-    /// Required safety tier for this tool, drawn directly from the mapping.
-    pub fn required_safety_tier(&self) -> SafetyTier {
-        self.mapping.safety_tier
-    }
-}
-
-#[async_trait]
-impl Tool for DoraToolBridge {
-    fn name(&self) -> &str {
-        &self.mapping.tool_name
-    }
-
-    fn description(&self) -> &str {
-        &self.mapping.description
-    }
-
-    fn input_schema(&self) -> serde_json::Value {
-        self.inner.input_schema()
-    }
-
-    fn tags(&self) -> &[&str] {
-        &["dora-bridge", "robot"]
-    }
-
-    /// Codex round-2 P2: anything that can actuate the robot must NOT run in
-    /// the same parallel batch as another bridge call against the same Dora
-    /// runtime — overlapping motion commands are a real safety hazard. Only
-    /// `Observe`-tier tools (sensors, status queries) keep the default
-    /// parallel-friendly `Safe` class; every higher tier is `Exclusive` so
-    /// the agent's batch dispatcher serialises them.
-    fn concurrency_class(&self) -> ConcurrencyClass {
-        if let Some(explicit) = self.mapping.concurrency_class {
-            return explicit;
-        }
-        match self.mapping.safety_tier {
-            SafetyTier::Observe => ConcurrencyClass::Safe,
-            SafetyTier::SafeMotion | SafetyTier::FullActuation | SafetyTier::EmergencyOverride => {
-                ConcurrencyClass::Exclusive
-            }
-        }
-    }
-
-    async fn execute(&self, args: &serde_json::Value) -> eyre::Result<ToolResult> {
-        self.inner.execute(args).await
-    }
-}
-
-/// Load tool mappings from a [`BridgeConfig`], create bridge tools, AND
-/// register each tool with the global [`RobotToolRegistry`] at its declared
-/// tier so the existing `group:robot:<tier>` `ToolPolicy` machinery sees
-/// the bridge tools. Without this registration the `safety_tier` field is
-/// decorative — group-based allow/deny silently misses every dora tool.
-///
-/// Idempotent: re-loading the same config replaces prior tier mappings for
-/// the same tool name (see [`RobotToolRegistry::insert`]).
-///
-/// Returns an error if any mapping has an invalid `bridge_base_url`
-/// (non-loopback / malformed) — the caller decides whether to abort.
-pub fn load_bridges(config: &BridgeConfig) -> eyre::Result<Vec<DoraToolBridge>> {
-    let bridges: Vec<DoraToolBridge> = config
-        .mappings
-        .iter()
-        .map(|m| DoraToolBridge::new(m.clone()))
-        .collect::<eyre::Result<Vec<_>>>()?;
-    robot_groups::with_registry_mut(|reg: &mut RobotToolRegistry| {
-        for bridge in &bridges {
-            reg.insert(bridge.mapping.tool_name.clone(), bridge.mapping.safety_tier);
-        }
-    });
-    Ok(bridges)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn sample_mapping() -> DoraToolMapping {
-        DoraToolMapping {
-            tool_name: "navigate_to".to_string(),
-            description: "Navigate robot to a waypoint".to_string(),
-            bridge_base_url: "http://127.0.0.1:8765".to_string(),
-            safety_tier: SafetyTier::SafeMotion,
-            concurrency_class: None,
-            input_schema: None,
-            timeout_secs: 60,
-            dora_node_id: None,
-            dora_output_id: None,
-        }
-    }
-
-    #[test]
-    fn should_expose_correct_tool_name() {
-        let bridge = DoraToolBridge::new(sample_mapping()).unwrap();
-        assert_eq!(bridge.name(), "navigate_to");
-    }
-
-    #[test]
-    fn should_expose_correct_description() {
-        let bridge = DoraToolBridge::new(sample_mapping()).unwrap();
-        assert_eq!(bridge.description(), "Navigate robot to a waypoint");
-    }
-
-    #[test]
-    fn should_include_dora_tags() {
-        let bridge = DoraToolBridge::new(sample_mapping()).unwrap();
-        let tags = bridge.tags();
-        assert!(tags.contains(&"dora-bridge"));
-        assert!(tags.contains(&"robot"));
-    }
-
-    #[test]
-    fn should_expose_declared_safety_tier() {
-        let bridge = DoraToolBridge::new(sample_mapping()).unwrap();
-        assert_eq!(bridge.required_safety_tier(), SafetyTier::SafeMotion);
-    }
-
-    #[test]
-    fn concurrency_class_override_supersedes_tier_default() {
-        let mut mapping = sample_mapping();
-        // sample_mapping is SafeMotion → would default to Exclusive.
-        mapping.concurrency_class = Some(ConcurrencyClass::Safe);
-        let bridge = DoraToolBridge::new(mapping).unwrap();
-        assert!(matches!(bridge.concurrency_class(), ConcurrencyClass::Safe));
-    }
-
-    #[test]
-    fn concurrency_class_falls_back_to_tier_default_when_unset() {
-        let bridge = DoraToolBridge::new(sample_mapping()).unwrap();
-        // SafeMotion → Exclusive.
-        assert!(matches!(
-            bridge.concurrency_class(),
-            ConcurrencyClass::Exclusive
-        ));
-    }
-
-    #[test]
-    fn should_fail_to_parse_unknown_safety_tier_string() {
-        // Codex round-1 P1: previously an unknown / misspelled tier silently
-        // collapsed to `Observe` (the LEAST restrictive), so a typo on a
-        // high-risk tool would slip past observe-only allow lists. Strict
-        // parsing now rejects the config.
-        let json = r#"{
-            "mappings": [{
-                "tool_name": "rogue",
-                "description": "should not load",
-                "bridge_base_url": "http://127.0.0.1:8765",
-                "safety_tier": "FullActuation",
-                "timeout_secs": 1
-            }]
-        }"#;
-        let result = BridgeConfig::from_json(json);
-        assert!(result.is_err(), "expected unknown-tier rejection, got Ok");
-    }
-
-    #[test]
-    fn should_register_bridges_in_robot_tool_registry_at_declared_tier() {
-        // Codex round-1 P2: the bridge previously carried `safety_tier` but
-        // never enrolled the tool in `RobotToolRegistry`, so the existing
-        // `group:robot:<tier>` ToolPolicy machinery had nothing to evaluate
-        // against. `load_bridges` now wires both halves.
-        let mut high = sample_mapping();
-        high.tool_name = "dora_register_high_actuation".to_string();
-        high.safety_tier = SafetyTier::FullActuation;
-
-        let mut low = sample_mapping();
-        low.tool_name = "dora_register_observe".to_string();
-        low.safety_tier = SafetyTier::Observe;
-
-        let config = BridgeConfig {
-            description: String::new(),
-            mappings: vec![high, low],
-        };
-        let bridges = load_bridges(&config).unwrap();
-        assert_eq!(bridges.len(), 2);
-
-        let snap = robot_groups::snapshot();
-        assert_eq!(
-            snap.tier_of("dora_register_high_actuation"),
-            Some(SafetyTier::FullActuation),
-        );
-        assert_eq!(
-            snap.tier_of("dora_register_observe"),
-            Some(SafetyTier::Observe),
-        );
-    }
-
-    #[tokio::test]
-    async fn execute_forwards_to_bridge_via_http() {
-        use wiremock::matchers::{body_json, method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        let server = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/tools/robot.heartbeat"))
-            .and(body_json(serde_json::json!({"args": {}})))
-            .respond_with(ResponseTemplate::new(200).set_body_json(serde_json::json!({
-                "ok": true, "code": "0", "msg": "", "data": {"ok": true, "deadline_ms": 0}
-            })))
-            .mount(&server)
-            .await;
-
-        let mapping = DoraToolMapping {
-            tool_name: "robot.heartbeat".into(),
-            description: "heartbeat".into(),
-            bridge_base_url: server.uri(),
-            safety_tier: SafetyTier::Observe,
-            concurrency_class: None,
-            timeout_secs: 5,
-            input_schema: None,
-            dora_node_id: None,
-            dora_output_id: None,
-        };
-        let bridge = DoraToolBridge::new(mapping).unwrap();
-        let result = bridge.execute(&serde_json::json!({})).await.unwrap();
-        assert!(result.success);
-        assert!(result.output.contains("deadline_ms"));
-    }
-}
+pub use octos_agent::tools::dora_bridge::{
+    BridgeConfig, DoraToolBridge, DoraToolMapping, default_timeout, load_bridges,
+};
